+++ lib/country_picker_cupertino.dart
@@ -1,7 +1,7 @@
-import 'package:country_picker_plus/countries.dart';
-import 'package:country_picker_plus/country.dart';
-import 'package:country_picker_plus/utils/typedefs.dart';
-import 'package:country_picker_plus/utils/utils.dart';
+import 'package:ka4_country_picker_plus/countries.dart';
+import 'package:ka4_country_picker_plus/country.dart';
+import 'package:ka4_country_picker_plus/utils/typedefs.dart';
+import 'package:ka4_country_picker_plus/utils/utils.dart';
 import 'package:flutter/cupertino.dart';
 import 'dart:core';