+++ lib/utils/typedefs.dart
@@ -1,4 +1,4 @@
-import 'package:country_picker_plus/country.dart';
+import 'package:ka4_country_picker_plus/country.dart';
 import 'package:flutter/material.dart';
 
 /// Returns true when a country should be included in lists / dialogs